refactor(cli): enable atlas only in `instantiateMetro` for both `expo start` and `expo export`
